# hard/hw1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Handlers для создания/поиска user/task",
        "title": "Rest API для обработки users и tasks",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Залогиниться на сайте и получить id сессии",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Авторизация",
                "parameters": [
                    {
                        "description": "username и password",
                        "name": "User_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ObjectTypeResultHandlerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Зарегаться на сайте",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Регистрация",
                "parameters": [
                    {
                        "description": "username и password",
                        "name": "User_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ObjectTypeResultHandlerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/result/{task_id}": {
            "get": {
                "description": "Получить результат таски по task_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Таска"
                ],
                "summary": "Получить результат таски",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id который получен от сервера",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer {auth_token}",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ObjectTypeResultHandlerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/status/{task_id}": {
            "get": {
                "description": "Получить статус таски по task_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Таска"
                ],
                "summary": "Готовность таски",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id который получен от сервера",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer {auth_token}",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ObjectTypeResultHandlerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/task": {
            "post": {
                "description": "Отправить таску на обработку",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Таска"
                ],
                "summary": "Отправить таску на обработку",
                "parameters": [
                    {
                        "description": "Таска которую нужно решить в string формате",
                        "name": "task_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer {auth_token}",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ObjectTypeResultHandlerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.ObjectTypeResultHandlerResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        }
    }
}